--- tests/utils/__init__.py
@@ -1,1 +1,0 @@
-from .helper_funcs import delete_records, insert_full_recordset, delete_all_records
